v1.14.0
This release aims at burning down the user request list. The big ones are further improvements to the configurability of the app as well as admin tools support. Some quality of life and bug fixes are also included in this release.

**Changelog**
- Added a setting to enable predictive back (API 34 and up only). When enabled, the app will no longer require the back gesture twice to close. Instead you will see an animation (from the Android OS) if your next back gesture will exit the app.
- Added support for sharing to Summit for Lemmy. When text is shared to Summit for Lemmy, the app will launch the create post screen and auto populate the URL field if the text is a URL or the body field otherwise. When an image is shared to Summit for Lemmy, the app will automatically upload the image to Lemmy and populate the URL field with the URL for the image.
- Added title suggestions for posts. When a URL is specified, a title will be suggested if the URL page has a title.
- Added the option to change sort order on a user's profile page. This sort order is used for both the user's post and comments.
- Added the option to toggle auto-loading on/off when using Infinity. When auto-loading is disabled, a button is shown to load the next page when the end is reached.
- Added the option to show page numbers in the post list.
- Added basic admin tools. Admins will be able to perform all basic mod actions on all communities in their instance.
- Added the admin action to ban a user from the instance. There appears to be a server side bug (needs confirming) where the ban will work but the server will report an error.
- Added admin actions to purge communities/users/posts/comments.
- Fixed a bug where refreshing a post will not update all of the post and comment scores.
- Changed the default sort order of content when viewing a person's profile to "New". Ie. the newest comments/posts will appear first. Previously there was no order so the ordering was arbitrary.
- Loading a comment that doesn't exist will now correctly show that the comment is missing instead of continuing to show a "load more" button.
- Fixed a crash when using the built in screen shot taker to take a screenshot of an image embedded in a text post.

+++ app/src/main/java/com/idunnololz/summit/account/AccountDao.kt
@@ -23,7 +23,7 @@ abstract class AccountDao {
     abstract suspend fun getFirstAccount(): Account?
 
     @Query("SELECT * FROM account WHERE id = :accountId")
-    abstract suspend fun getAccountById(accountId: Int): Account?
+    abstract suspend fun getAccountById(accountId: Long): Account?
 
     @Insert(onConflict = OnConflictStrategy.IGNORE, entity = Account::class)
     abstract suspend fun insert(account: Account)
@@ -35,7 +35,7 @@ abstract class AccountDao {
     abstract suspend fun removeCurrent()
 
     @Query("UPDATE account set current = 1 where id = :accountId")
-    abstract suspend fun setCurrent(accountId: Int)
+    abstract suspend fun setCurrent(accountId: Long)
 
     @Delete(entity = Account::class)
     abstract suspend fun delete(account: Account)
@@ -44,7 +44,7 @@ abstract class AccountDao {
     abstract suspend fun count(): Int
 
     @Transaction
-    open suspend fun clearAndSetCurrent(accountId: Int) {
+    open suspend fun clearAndSetCurrent(accountId: Long) {
         removeCurrent()
         setCurrent(accountId)
     }
