+++ app/src/main/java/com/idunnololz/summit/api/dto/PersonId.kt
@@ -1,3 +1,3 @@
 package com.idunnololz.summit.api.dto
 
-typealias PersonId = Int
+typealias PersonId = Long
